Critical fix: table order was cronological.
The order of the table cards shown inline in the voting phase was the same as
the order in which the players dealt their cards. In particular, the
storyteller's card came always first (how hadn't we figured this out
before!?).

Now, the list of table cards is shuffled with a seed that
depends only on the particular round and game. This is so that 1.
players cannot memorize the permutation; but at the same time, 2. the
cards are not shuffled again if the player calls inline another time.

To avoid confusion, I have changed previous mentions of "game id" to
"chat id" when appropriate.