Saves all data in JSON.
It uses a modified python-telegram-bot's persistance class based on JSON
dictionaries. The main work was put into converting all custom classes
into dictionaries to be JSON serializable.

It currently does not read from the data files in data/. However,
foreseeing that, I have changed the argument of the classes so that we
could create them again from the JSON files.