Implement `Player` name formatting via __format__
Now the formatting of the player's name can be done with the help of the
`__format__` magic method. This means that, given a `spec` format
string, we can either call `format(player, spec)` or `f"{player:spec}"
to get the formatted name. See the method's documentation for the valid
spec strings.