Load/write every image from/to memory.
Loads the cairo surfaces for all card images into memory before
starting bot and uses them to construct the results. Furthermore, no
longer saves the results into tmp/, instead passing it from draw.py to
main.py through memory.

Advantages: Possible 20% increase in speed. May be a fluke and there is
effectively 0% increase. Did not do many tests.

Disadvantages: Slow to start bot and consumes 0.5 GB of memory. Both can
be avoided if we save the images to memory istead of the Cairo surfaces,
and there seems to be no speed penalty to doing that.

Verdict: Everything related to file handling is now in memory, so that's
not what is causing the slowdown. I think it's Cairo in some way, but I
haven't profiled it yet. Maybe shrinking the png images down could
improve speed without changing much of the code.